requirements.txt
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
cmdstanpy==1.2.0
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
google-analytics-data==0.17.2
google-api-core==2.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.0
google-api-python-client==2.34.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
gspread==5.12.0
gspread-dataframe==3.3.1
holidays==0.35
httplib2==0.22.0
idna==3.4
importlib-resources==6.1.0
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
prophet==1.1.5
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stanio==0.3.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.7
plotly